Initial functionality.
This commit implements the systemctl service, and the dbus
service. It adds commands for inhibiting/uninhibiting the
screensaver as well as a toggle command and a query command
for the current state.

The point of this is that the standard freedesktop commands for
inhibiting the screensaver will track the requesting process so
they can clear the wakelock when the process terminates. That
makes complete sense (you don't want a wakelock held forever
by a process that crashed) but that also means you need a
long-lived process to inhibit the screensaver manually, e.g.
with a keyboard shortcut. This commit, on top of adding the
daemon setup, implements this long-lived process.

Setting their favorite keyboard shortcuts in their desktop
environment is left to the user.